moved some types to types.go
